Custom Hook (useTickets) + Detail Button (no function yet).
